+++ DKBasicEngine 1.0/Core/Interfaces/IText.cs
@@ -2,6 +2,6 @@
 {
     interface IText
     {
-        double FontSize { get; set; }
+        float FontSize { get; set; }
     }
 }
